Added additional callback functions
Added callback functions that are called when the slider reaches the
first and last slides. Added cycles boolean to determine whether
slideshow loops back to start when reaching end. Added callback
functions that are called if cycle is set to false. This can be used,
for example, to load a new page at the end of a set of slides.